fix uint64 fields to NUMERIC(1000,0) + update batching

+++ db/migrations/000003_add_chain_details_to_node_info.up.sql
@@ -2,5 +2,5 @@
 ALTER TABLE node_info ADD COLUMN fork_id TEXT;
 ALTER TABLE node_info ADD COLUMN protocol_version INT;
 ALTER TABLE node_info ADD COLUMN head_hash TEXT;
-ALTER TABLE node_info ADD COLUMN network_id INT;
-ALTER TABLE node_info ADD COLUMN total_difficulty BIGINT;
+ALTER TABLE node_info ADD COLUMN network_id NUMERIC(1000, 0);
+ALTER TABLE node_info ADD COLUMN total_difficulty NUMERIC(1000, 0);
